Move extraPayload() functions
